🧹 Misc cleanup
- Was pointless to have one event for render and one for
  cancel. Reduced it to one while fixing a minor bug.
- Remove useless dereference and borrow. Not sure how it
  ended up like that.
- Moved around some code.
